Make ecef public
